Split up dice rolling and moving
Split the dice roll and movement into two methods, one which rolls the dice and returns the values, along with if they are the same, and one which takes in the dice values and moves the counter. I also added a method that rolls the dice and moves the counter, and repeats this if the dice rolls are the same.